remove create command
